+++ Sources/GanitWorkspaceUI/MainMenu.swift
@@ -57,6 +57,7 @@ import GanitEditorUI
   func setNumberFormat(_ sender: Any?)
   func toggleDigitGrouping(_ sender: Any?)
   func toggleLakhGrouping(_ sender: Any?)
+  func toggleDegrees(_ sender: Any?)
   func toggleMarkdownMode(_ sender: Any?)
   func setDollarCurrency(_ sender: Any?)
   func toggleAnswerSeparator(_ sender: Any?)
@@ -368,6 +369,9 @@ public enum MainMenu {
           item(
             localized("menu.groupInLakhs", "Group Digits in Lakhs"),
             #selector(WorkspaceCommands.toggleLakhGrouping(_:))),
+          item(
+            localized("menu.anglesInDegrees", "Angles in Degrees"),
+            #selector(WorkspaceCommands.toggleDegrees(_:))),
           .separator(),
           item(
             localized("menu.markdownMode", "Markdown Mode"),